pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hpc-submit"
version = "0.1.0"
description = "Submit SLURM job scripts to a remote HPC cluster via SSH"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
hpc-submit = "hpc_submit.cli:main"
